pyproject.toml
[project]
name = "devproxy"
version = "0.1.0"
description = "Local HTTPS development proxy with subdomain routing and request inspection"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Charlie Kolb" }]
keywords = ["proxy", "https", "development", "mitmproxy", "ssl", "tls"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.15.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "PyYAML>=6.0.0",
    "mitmproxy>=11.0.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
devproxy = "devproxy.cli.main:app"

[project.urls]
Homepage = "https://github.com/charliek/devproxy"
Documentation = "https://github.com/charliek/devproxy#readme"
Repository = "https://github.com/charliek/devproxy"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devproxy"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["devproxy"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
